+++ RELEASES.md
@@ -48,6 +48,7 @@ Example usage change:
   - The `axis` module has been merged into the `geometry` module
   - The debug module is no longer public. The `print_tree` function is now accesible under `util`.
   - All types from the `node`, `data`, `layout`, `error` and `cache` modules have been moved to the  the `tree` module.
+- Fixed misspelling: `RunMode::PeformLayout` renamed into `RunMode::PerformLayout` (added missing `r`).
 
 ### Fixes